[Cleanup] [MSVC Compiler Warnings] [Several type-changes. No functional change.]
- ident_c.h: ident(pos) will now return uint8_t instead of MASK_TYPE
- target_extension_socket: using INVALID_SOCKET instead of -1 now correctly for SOCKET_TYPE.
- multisend: size is now 16bit instead of 32bit.
- scheduler: getExecTime now changed to i32 isntead of ui16.

Done by Bernd Jäkel
Reviewed/Adapted by Martin Wodok
Post-Review by Fischer/Jäkel

+++ library/xgpl_src/IsoAgLib/scheduler/impl/scheduler_c.h
@@ -160,7 +160,7 @@ public:
     * refine time schedule within execution
     * @return average execution time in [msec] (off complete performed runs)
     */
-  uint16_t getExecTime( void ) const { return mi32_averageExecTime;}
+  int32_t getExecTime( void ) const { return mi32_averageExecTime;}
 
   /**
     * if a very imprtant IRQ event forces stop of Scheduler_c::timeEvent AS SOON AS POSSIBLE
